requirements.txt
flask
yt-dlp
faster-whisper
ffmpeg-python
gunicorn
python-dotenv
werkzeug
beautifulsoup4